requirements.txt
# grid-detection-agent
pdf2image==1.17.0
Pillow
requests

# gltf exporter
trimesh
numpy

# yolo detection agents
ultralytics>=8.3.0
torchvision>=0.16.0
pymupdf>=1.24.0

# type resolution agents
pytesseract>=0.3.10
opencv-python-headless>=4.9.0
